Add files via upload
This registration form has been made in C++, includes basic functionality of Login and Registering a user. It uses the basic concepts of file handling and matches the input data with the stored data in .txt file. The code is easy to understand and just needed to be pasted in the field of requirement.